requirements.txt
python-dotenv
azure.core
langchain==0.0.240
openai==0.27.8
pandas
tiktoken
PyPDF2
matplotlib
plotly
scipy
scikit-learn
azure.kusto.data
azure-synapse-spark
pyspark
redis
pypdf
pymongo
faiss-cpu
azure-kusto.data
pymssql
unstructured
youtube-transcript-api
pytube
wikipedia
azure.cosmos
pymongo
ipywidgets
plotly.express
nbformat>=4.2.0
streamlit
cohere
azure-ai-formrecognizer
azure.core
azure-cognitiveservices-vision-computervision
Pillow
azure-ai-vision
azure-cognitiveservices-speech